core: add self sub-command and adjust install step
1. add xlings self [command]
2. auto init xlings for first run - [global install]
3. update install script
