+++ core/common.lua
@@ -228,127 +228,6 @@ function xlings_path_format(path)
     return path
 end
 
-function xlings_install()
-
-    xlings_uninstall() -- TODO: avoid delete mdbook?
-
-    cprint("[xlings]: install xlings to %s", platform.get_config_info().install_dir)
-
-    local install_dir = platform.get_config_info().install_dir
-    -- cp xliings to install_dir and force overwrite
-    os.cp(platform.get_config_info().sourcedir, install_dir, {force = true})
-
-    -- create rcachedir
-    local rcachedir = platform.get_config_info().rcachedir
-
-    -- TODO: check cache data is valid
-    if not os.isfile(path.join(rcachedir, "xlings.json")) then
-        cprint("[xlings]: create rcachedir %s", rcachedir)
-        os.mkdir(platform.get_config_info().rcachedir)
-
-        -- create bin dir
-        local bindir = platform.get_config_info().bindir
-        cprint("[xlings]: create bindir %s", bindir)
-        os.cp(path.join(install_dir, "bin"), bindir, {force = true})
-
-        -- copy profile to rcachedir
-        cprint("[xlings]: copy profile to rcachedir...")
-        os.cp(path.join(install_dir, "config", "shell", "xlings-profile.*"), rcachedir, {force = true})
-
-        -- copy xlings config file to rcachedir
-        cprint("[xlings]: copy xlings config file to rcachedir...")
-        os.cp(path.join(install_dir, "config", "xlings.json"), rcachedir, {force = true})
-    else
-        cprint("[xlings]: use local cache data - %s", rcachedir)
-    end
-
-    -- add bin to linux bashrc and windows's path env
-    cprint("[xlings]: config system environment...")
-
-    if is_host("linux") then
-        local source_cmd_template = "\ntest -f %s && source %s"
-        local posix_profile = path.join(rcachedir, "xlings-profile.sh")
-        local fish_profile = path.join(rcachedir, "xlings-profile.fish")
-        local shells = {
-            bash = {
-                profile = ".bashrc",
-                command = format(source_cmd_template, posix_profile, posix_profile)
-            },
-            zsh = {
-                profile = ".zshrc",
-                command = format(source_cmd_template, posix_profile, posix_profile)
-            },
-            fish = {
-                profile = ".config/fish/config.fish",
-                command = format("\ntest -f %s; and source %s", fish_profile, fish_profile)
-            }
-        }
-
-        function config_shell(shell)
-            local profile = path.join(os.getenv("HOME"), shells[shell].profile)
-            local command = shells[shell].command
-            if os.isfile(profile) then
-                cprint("[xlings]: config [%s] shell - %s", shell, profile)
-                local profile_content = io.readfile(profile)
-                if not string.find(profile_content, command, 1, true) then
-                    xlings_file_append(profile, command)
-                else
-                    cprint("[xlings]: [%s] already configed", shell)
-                end
-            end
-        end
-
-        for shell, _ in pairs(shells) do
-            config_shell(shell)
-        end
-    else
-        local path_env = os.getenv("PATH")
-        if not string.find(path_env, install_dir, 1, true) then
-
-            path_env = path_env .. ";" .. install_dir
-            -- os.setenv("PATH", path_env) -- only tmp, move to install.win.bat
-            -- os.exec("setx PATH " .. path_env)
-        end
-    end
-end
-
-function xlings_uninstall()
-    local install_dir = platform.get_config_info().install_dir
-    local rcachedir = platform.get_config_info().rcachedir
-    try
-    {
-        function()
-            os.rm(install_dir)
-            cprint("[xlings]: remove %s - ok", install_dir)
-            -- check rcachedir not empty by xlings.json
-            if os.isfile(path.join(rcachedir, "xlings.json")) then
-                cprint("${cyan blink}-> ${clear}${yellow}delete local cache data?(y/n)")
-                local confirm = io.read()
-                if confirm == "y" then
-                    cprint("[xlings]: remove %s - ok", rcachedir)
-                    os.rm(rcachedir)
-                end
-            end
-        end,
-        catch
-        {
-            function (e)
-                -- TODO: error: cannot remove directory C:\Users\Public\xlings Unknown Error (145)
-                cprint("[xlings]: xlings_uninstall: " .. e)
-            end
-        }
-    }
-
-    cprint("[xlings]: xlings uninstalled - ok")
-
-end
-
-function xlings_update()
-    cprint("[xlings]: try to update xlings...")
-    local install_dir = platform.get_config_info().install_dir
-    git.pull({remote = "origin", tags = true, branch = "main", repodir = install_dir})
-end
-
 -- download repo from git_url to destination directory
 function xlings_download_repo(git_url, dest_dir)
     local repo_name = git_url:match("([^/]+)%.git$")
